Integrate back implementation logic to the endpoints
