--- crates/starknet-devnet-server/src/api/endpoints_impl/mod.rs
@@ -1,10 +1,0 @@
-/// Accounts
-pub mod accounts;
-/// Dumping and loading
-pub mod dump_load;
-/// Mint token - Local faucet
-pub mod mint_token;
-/// Postman
-pub mod postman;
-/// Time
-pub mod time;
